Make all tests known to pytest
This allows seeing the following tests individually:
```
visidata/fuzzymatch.py::test_fuzzymatch
visidata/features/slide.py::test_slide_keycol_1
visidata/features/slide.py::test_slide_leftmost
```
instead of all being hidden away in a single test in `test_features.py`.
